requirements.txt
playwright==1.42.0
playwright-stealth==1.0.6
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1